+++ app/controllers/esign_templates_controller.py
@@ -20,10 +20,14 @@ class EsignTemplatesController(BaseController, Authenticatable):
         return au.account_id
 
     def index(self):
-        r = EsignTemplateService(self.db).list_templates(self._account_id())
+        account_id = self._account_id()
+        self.require_permission("esign", "view", account_id=account_id)
+        r = EsignTemplateService(self.db).list_templates(account_id)
         return self.render_json(r["data"])
 
     def show(self):
+        account_id = self._account_id()
+        self.require_permission("esign", "view", account_id=account_id)
         tid = int(self.request.path_params["id"])
         r = EsignTemplateService(self.db).get_template(self._account_id(), tid)
         if not r["ok"]:
@@ -31,6 +35,7 @@ class EsignTemplatesController(BaseController, Authenticatable):
         return self.render_json(r["data"])
 
     async def create(self):
+        self.require_permission("esign", "manage", account_id=self._account_id())
         data = await self._get_body_json()
         r = EsignTemplateService(self.db).create_template(self._account_id(), data)
         if not r["ok"]:
@@ -38,6 +43,7 @@ class EsignTemplatesController(BaseController, Authenticatable):
         return self.render_json(r["data"], status=201)
 
     async def update(self):
+        self.require_permission("esign", "manage", account_id=self._account_id())
         tid = int(self.request.path_params["id"])
         data = await self._get_body_json()
         r = EsignTemplateService(self.db).update_template(self._account_id(), tid, data)
@@ -46,6 +52,7 @@ class EsignTemplatesController(BaseController, Authenticatable):
         return self.render_json(r["data"])
 
     def destroy(self):
+        self.require_permission("esign", "manage", account_id=self._account_id())
         tid = int(self.request.path_params["id"])
         r = EsignTemplateService(self.db).delete_template(self._account_id(), tid)
         if not r["ok"]:
